Ficha medica, agregar, borrar, modificar tablas

+++ delete.php
@@ -4,7 +4,14 @@
     $id = $_GET['id'];
     $table = $_GET['table'];
 
-    if ($conn -> query("DELETE FROM $table WHERE `id`= $id")) {    
+    if ($table == 'notificaciones') {
+        if ($conn->query("DELETE FROM $table WHERE `id`= $id")) {    
+            return;
+        } else {
+            echo $conn->error;
+        }
+    }
+    if ($conn->query("DELETE FROM $table WHERE `id`= $id")) {    
         header('Location: ./dashboard');
     } else {
         echo $conn->error;
